apply ALL requested changes
