--- Libraries/Starlight.Common/SdkDefaults.cs
@@ -1,16 +1,0 @@
-namespace Starlight.Common;
-
-public static class SdkDefaults
-{
-    public const string DefaultCountryCode = "US";
-    public const string MaPassportLanguage = "en-us";
-    public const string ApiLogLevel = "DEBUG";
-    public const string ShieldIdentity = "I_IDENTITY";
-    public const string ShieldSceneNormal = "S_NORMAL";
-    public const string ShieldSceneAccount = "S_ACCOUNT";
-    public const string ShieldIgnoreVersions = "2.6.0";
-    public const string DefaultComboId = "0";
-    public const string GuardianAgeLimit = "14";
-    public const string DefaultMid = "18w9wecbdl_hy";
-    public const string ZeroTimestamp = "0";
-}
